@@ -5,7 +5,7 @@
   "description": "React.js Google Maps integration component",
   "main": "lib/index.js",
   "scripts": {
-    "build": "jsx --harmony ./src ./lib",
+    "build": "rm -rf ./lib && babel --stage=1 ./src --out-dir ./lib",
     "build_client": "webpack -p --config ./client/webpack.config.js",
     "test": "jest",
     "dev_install": "npm install && npm dedupe && bower install",
@@ -51,7 +51,12 @@
   },
   "homepage": "https://github.com/tomchentw/react-google-maps",
   "devDependencies": {
+    "babel": "^5.1.13",
+    "babel-core": "^5.1.13",
+    "babel-jest": "^5.0.1",
+    "babel-loader": "^5.0.0",
     "bower": "^1.3.9",
+    "classnames": "^1.2.1",
     "css-loader": "^0.9.0",
     "expose-loader": "^0.5.3",
     "file-loader": "^0.8.1",
@@ -60,10 +65,9 @@
     "jquery": "^2.1.1",
     "jsx-loader": "^0.12.0",
     "raw-loader": "^0.5.1",
-    "react-hot-loader": "^0.5.0",
-    "react-prism": "^1.0.0",
-    "react-toastr": "^1.1.2",
-    "react-tools": "^0.12.0",
+    "react-hot-loader": "^1.2.5",
+    "react-prism": "^1.3.0",
+    "react-toastr": "^1.2.0",
     "sass-loader": "^0.3.0",
     "style-loader": "^0.8.1",
     "tomchentw-npm-dev": "^1.1.0",
@@ -71,11 +75,13 @@
     "webpack-dev-server": "^1.6.5"
   },
   "dependencies": {
-    "react": "^0.12.0",
-    "deep-equal": "^0.2.1"
+    "react": "^0.13.0"
+  },
+  "peerDependencies": {
+    "react": "^0.13.0"
   },
   "jest": {
-    "scriptPreprocessor": "<rootDir>/test/__preprocessor__.js",
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
     "testPathDirs": [
       "<rootDir>/src"
     ],
