@@ -1,43 +1,61 @@
-"use strict";
-var React = require("react/addons"),
+import React from "react/addons";
+import {GoogleMaps, InfoWindow, Marker} from "react-google-maps";
 
-    {GoogleMapsMixin, Map, InfoWindow} = require("react-google-maps"),
-    GettingProperties;
 /*
  * https://developers.google.com/maps/documentation/javascript/examples/event-properties
  */
-GettingProperties = React.createClass({
-  displayName: "GettingProperties",
-
-  mixins: [require("../../ReactFutureMixin"), GoogleMapsMixin],
+const GettingProperties = React.createClass({
 
   getInitialState () {
     return {
+      zoomLevel: 4,
       content: "Change the zoom level", 
     };
   },
 
   _handle_zoom_changed () {
-    var zoomLevel = this.refs.map.getZoom();
-    this.setState({
-      content: `Zoom: ${ zoomLevel }`,
-    });
+    const zoomLevel = this.refs.map.getZoom();
+    if (zoomLevel !== this.state.zoomLevel) {
+      // Notice: Check zoomLevel equality here,
+      // or it will fire zoom_changed event infinitely
+      this.setState({
+        zoomLevel,
+        content: `Zoom: ${ zoomLevel }`,
+      });
+    }
   },
 
-  _render (props, state) {
-    var myLatLng = new google.maps.LatLng(-25.363882, 131.044922);
+  render () {
+    const {props, state} = this,
+          {googleMapsApi, ...otherProps} = props,
+          myLatLng = new google.maps.LatLng(-25.363882, 131.044922);
 
-    return <div style={{height: "100%"}} {...props}>
-      <Map ref="map" style={{height: "100%"}} zoom={4} center={myLatLng} onZoomChanged={this._handle_zoom_changed} />
-      <InfoWindow position={myLatLng} content={state.content} />
-    </div>;
+    return (
+      <GoogleMaps containerProps={{
+          ...otherProps,
+          style: {
+            height: "100%",
+          },
+        }} mapProps={{
+          style: {
+            height: "100%",
+          },
+        }}
+        ref="map"
+        googleMapsApi={googleMapsApi}
+        zoom={state.zoomLevel}
+        center={myLatLng}
+        onZoomChanged={this._handle_zoom_changed}>
+        <InfoWindow position={myLatLng} content={state.content} />
+      </GoogleMaps>
+    );
   }
 });
 
-module.exports = React.createClass({
-  mixins: [require("../../ReactFutureMixin")],
-
-  _render (props, state) {
-    return <GettingProperties googleMapsApi={google.maps} {...props} />;
+export default React.createClass({
+  render () {
+    return (
+      <GettingProperties googleMapsApi={google.maps} {...this.props} />
+    );
   }
 });
