@@ -8,7 +8,7 @@ var Path = require("path"),
     webpackConfig,
 
     IS_PRODUCTION = "production" === process.env.NODE_ENV,
-    JSX_WITH_HOT_LOEADERS = ["react-hot-loader", "jsx-loader?harmony"],
+    JSX_WITH_HOT_LOEADERS = ["react-hot-loader", "babel-loader?stage=1"],
     CSS_LOADER = "style-loader!css-loader?root=../",
     SCSS_LOADER = CSS_LOADER + "!sass-loader?" + JSON.stringify({
       includePaths: [
@@ -33,8 +33,7 @@ webpackConfig = module.exports = {
   },
   module: {
     loaders: [
-      { test: require.resolve("react/addons"), loader: "expose-loader?React" },
-      { test: /\.js(x?)$/, loaders: JSX_WITH_HOT_LOEADERS },
+      { test: /\.js(x?)$/, exclude: /node_modules/, loaders: JSX_WITH_HOT_LOEADERS },
       { test: /\.jpg$/, loader: "file-loader" },
       { test: /\.css$/, loader: CSS_LOADER },
       { test: /\.scss$/, loader: SCSS_LOADER },
